pyproject.toml
[project]
name = "bnbagent"
version = "0.1.0"
description = "Modular Python SDK for on-chain AI agents on BNB Chain"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
keywords = [
    "erc-8004",
    "erc-8183",
    "apex",
    "blockchain",
    "ethereum",
    "agent",
    "web3",
    "bsc",
    "binance-smart-chain",
    "sdk",
    "modular",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet",
]
dependencies = [
    "web3>=6.15.0",
    "eth-account>=0.10.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.urls]
Homepage = "https://github.com/bnb-chain/bnbagent-sdk"
Documentation = "https://github.com/bnb-chain/bnbagent-sdk#readme"
Repository = "https://github.com/bnb-chain/bnbagent-sdk"
Issues = "https://github.com/bnb-chain/bnbagent-sdk/issues"

[project.optional-dependencies]
server = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
]
ipfs = [
    "httpx>=0.25.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.11.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.25.0",
    "fastapi>=0.104.0",
]

[project.entry-points."bnbagent.modules"]
erc8004 = "bnbagent.erc8004:create_module"
apex = "bnbagent.apex:create_module"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
exclude = [
    ".claude/",
    ".github/",
    ".venv/",
    "_docs/",
    "dist/",
]

[tool.hatch.build.targets.wheel]
packages = ["bnbagent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
